packages/integrations/google-adk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "occ-google-adk"
version = "0.1.0"
description = "OCC cryptographic proof signing for Google ADK"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Mike Argento" },
]
keywords = ["occ", "proof", "ed25519", "google-adk", "agents", "cryptography"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "PyNaCl>=1.5.0",
    "google-adk>=0.1.0",
]

[project.urls]
Homepage = "https://proofstudio.xyz"
Repository = "https://github.com/mikeargento/occ"

[tool.hatch.build.targets.wheel]
packages = ["src/occ_google_adk"]
